+++ src/main/java/api/simplequotes/application/repositores/QuotationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import api.simplequotes.application.models.Quotation;
 
-public interface QuotationRepository extends CrudRepository<Quotation, Integer> {
+public interface QuotationRepository extends CrudRepository<Quotation, Long> {
   boolean existsByQuotationContent(String quotationContent);
 }
